[add]添加历史操作记录系统，实现部门与人员记录的每一次编辑，删除，新增，新增member_log与department_log表以及后端相关逻辑，同时push在主分支，完成手动合并

+++ backend/src/main/java/com/example/backend/entity/SavingAccount.java
@@ -2,6 +2,8 @@ package com.example.backend.entity;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class SavingAccount {
     private Integer account_id;
@@ -10,5 +12,5 @@ public class SavingAccount {
     private String password;
     private Integer remaining;
     private String open_date;
-    private Float interest_rate;
+    private BigDecimal interest_rate;
 }
